sql: remove old EXPLAIN code
Release note: None
