+++ pkg/sql/plan_columns.go
@@ -63,8 +63,6 @@ func getPlanColumns(plan planNode, mut bool) sqlbase.ResultColumns {
 		return n.columns
 	case *virtualTableNode:
 		return n.columns
-	case *explainPlanNode:
-		return n.run.results.columns
 	case *explainNewPlanNode:
 		return n.columns
 	case *windowNode:
