+++ pkg/sql/walk.go
@@ -11,25 +11,10 @@
 package sql
 
 import (
-	"bytes"
 	"context"
-	"fmt"
 	"reflect"
-	"strings"
 
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
-	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/errors"
-)
-
-type observeVerbosity int
-
-const (
-	observeMetadata observeVerbosity = iota
-	observeAlways
 )
 
 // planObserver is the interface to implement by components that need
@@ -45,16 +30,6 @@ type planObserver struct {
 	// enterNode is invoked upon entering a tree node. It can return false to
 	// stop the recursion at this node.
 	enterNode func(ctx context.Context, nodeName string, plan planNode) (bool, error)
-
-	// expr is invoked for each expression field in each node.
-	expr func(verbosity observeVerbosity, nodeName, fieldName string, n int, expr tree.Expr)
-
-	// spans is invoked for spans embedded in each node. hardLimitSet indicates
-	// whether the node will "touch" a limited number of rows.
-	spans func(nodeName, fieldName string, index *descpb.IndexDescriptor, spans []roachpb.Span, hardLimitSet bool)
-
-	// attr is invoked for non-expression metadata in each node.
-	attr func(nodeName, fieldName, attr string)
 
 	// leaveNode is invoked upon leaving a tree node.
 	leaveNode func(nodeName string, plan planNode) error
@@ -151,387 +126,56 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 
 	switch n := plan.(type) {
 	case *valuesNode:
-		if v.observer.attr != nil {
-			numRows := len(n.tuples)
-			if n.rows != nil {
-				numRows = n.rows.Len()
-			}
-			v.observer.attr(name, "size", formatValuesSize(numRows, len(n.columns)))
-		}
-
-		if v.observer.expr != nil {
-			v.metadataTuples(name, n.tuples)
-		}
-
 	case *scanNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", formatTable(n.desc, n.index))
-			if n.noIndexJoin {
-				v.observer.attr(name, "hint", "no index join")
-			}
-			if n.specifiedIndex != nil {
-				v.observer.attr(name, "hint", fmt.Sprintf("force index @%s", n.specifiedIndex.Name))
-			}
-		}
-		if v.observer.spans != nil {
-			v.observer.spans(name, "spans", n.index, n.spans, n.hardLimit != 0)
-		}
-		if v.observer.attr != nil {
-			// Only print out "parallel" when it makes sense. i.e. don't print if
-			// we know we will get only one result from the scan. There are cases
-			// in which "parallel" will be printed out even though the spans cover
-			// a single range, but there is nothing we can do about that.
-			if n.parallelize {
-				v.observer.attr(name, "parallel", "")
-			}
-			if n.hardLimit > 0 {
-				v.observer.attr(name, "limit", fmt.Sprintf("%d", n.hardLimit))
-			}
-			if n.lockingStrength != descpb.ScanLockingStrength_FOR_NONE {
-				v.observer.attr(name, "locking strength", n.lockingStrength.PrettyString())
-			}
-			if n.lockingWaitPolicy != descpb.ScanLockingWaitPolicy_BLOCK {
-				v.observer.attr(name, "locking wait policy", n.lockingWaitPolicy.PrettyString())
-			}
-		}
 
 	case *filterNode:
-		if v.observer.expr != nil {
-			v.expr(name, "filter", -1, n.filter)
-		}
 		n.source.plan = v.visit(n.source.plan)
 
 	case *renderNode:
-		if v.observer.expr != nil {
-			for i, r := range n.render {
-				v.metadataExpr(name, "render", i, r)
-			}
-		}
 		n.source.plan = v.visit(n.source.plan)
 
 	case *indexJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", fmt.Sprintf("%s@%s", n.table.desc.Name, n.table.index.Name))
-			inputCols := planColumns(n.input)
-			cols := make([]string, len(n.keyCols))
-			for i, c := range n.keyCols {
-				cols[i] = inputCols[c].Name
-			}
-			v.observer.attr(name, "key columns", strings.Join(cols, ", "))
-		}
 		n.input = v.visit(n.input)
 
 	case *lookupJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", fmt.Sprintf("%s@%s", n.table.desc.Name, n.table.index.Name))
-			v.observer.attr(name, "type", joinTypeStr(n.joinType))
-		}
-		var b bytes.Buffer
-		b.WriteByte('(')
-		inputCols := planColumns(n.input)
-		for i, c := range n.eqCols {
-			if i > 0 {
-				b.WriteString(", ")
-			}
-			b.WriteString(inputCols[c].Name)
-		}
-		b.WriteString(") = (")
-		for i := range n.eqCols {
-			if i > 0 {
-				b.WriteString(", ")
-			}
-			if i < len(n.table.index.ColumnNames) {
-				b.WriteString(n.table.index.ColumnNames[i])
-			} else {
-				id := n.table.index.ExtraColumnIDs[i-len(n.table.index.ColumnNames)]
-				col, err := n.table.desc.FindColumnByID(id)
-				if err != nil {
-					fmt.Fprintf(&b, "<error: %v>", err)
-				} else {
-					b.WriteString(col.Name)
-				}
-			}
-		}
-		b.WriteByte(')')
-		v.observer.attr(name, "equality", b.String())
-		if n.eqColsAreKey {
-			v.observer.attr(name, "equality cols are key", "")
-		}
-		if n.CanParallelize() {
-			v.observer.attr(name, "parallel", "")
-		}
-		if v.observer.expr != nil && n.onCond != nil && n.onCond != tree.DBoolTrue {
-			v.expr(name, "pred", -1, n.onCond)
-		}
 		n.input = v.visit(n.input)
 
 	case *vTableLookupJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", fmt.Sprintf("%s@%s", n.table.Name, n.index.Name))
-			v.observer.attr(name, "type", joinTypeStr(n.joinType))
-			var b bytes.Buffer
-			b.WriteByte('(')
-			inputCols := planColumns(n.input)
-			b.WriteString(inputCols[n.eqCol].Name)
-			b.WriteString(") = (")
-			b.WriteString(n.index.ColumnNames[0])
-			b.WriteByte(')')
-			v.observer.attr(name, "equality", b.String())
-		}
-		if v.observer.expr != nil && n.pred.onCond != nil && n.pred.onCond != tree.DBoolTrue {
-			v.expr(name, "pred", -1, n.pred.onCond)
-		}
 		n.input = v.visit(n.input)
 
 	case *zigzagJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "type", joinTypeStr(descpb.InnerJoin))
-			if v.observer.expr != nil && n.onCond != nil && n.onCond != tree.DBoolTrue {
-				v.expr(name, "pred", -1, n.onCond)
-			}
-			for _, side := range n.sides {
-				v.visitConcrete(side.scan)
-				if side.fixedVals != nil {
-					description := fmt.Sprintf(
-						"%d column%s",
-						len(side.fixedVals.columns),
-						util.Pluralize(int64(len(side.fixedVals.columns))),
-					)
-					v.observer.attr(name, "fixedvals", description)
-				}
-			}
-		}
 
 	case *applyJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "type", joinTypeStr(n.joinType))
-		}
-		if v.observer.expr != nil {
-			v.expr(name, "pred", -1, n.pred.onCond)
-		}
 		n.input.plan = v.visit(n.input.plan)
 
 	case *joinNode:
-		if v.observer.attr != nil {
-			jType := joinTypeStr(n.joinType)
-			if n.joinType == descpb.InnerJoin && len(n.pred.leftColNames) == 0 && n.pred.onCond == nil {
-				jType = "cross"
-			}
-			v.observer.attr(name, "type", jType)
-
-			if len(n.pred.leftColNames) > 0 {
-				f := tree.NewFmtCtx(tree.FmtSimple)
-				f.WriteByte('(')
-				f.FormatNode(&n.pred.leftColNames)
-				f.WriteString(") = (")
-				f.FormatNode(&n.pred.rightColNames)
-				f.WriteByte(')')
-				v.observer.attr(name, "equality", f.CloseAndGetString())
-				if n.pred.leftEqKey {
-					v.observer.attr(name, "left cols are key", "")
-				}
-				if n.pred.rightEqKey {
-					v.observer.attr(name, "right cols are key", "")
-				}
-			}
-			if len(n.mergeJoinOrdering) > 0 {
-				// The ordering refers to equality columns
-				eqCols := make(sqlbase.ResultColumns, len(n.pred.leftEqualityIndices))
-				for i := range eqCols {
-					eqCols[i].Name = fmt.Sprintf("(%s=%s)", n.pred.leftColNames[i], n.pred.rightColNames[i])
-				}
-				v.observer.attr(name, "mergeJoinOrder", n.mergeJoinOrdering.String(eqCols))
-			}
-		}
-		if v.observer.expr != nil {
-			v.expr(name, "pred", -1, n.pred.onCond)
-		}
 		n.left.plan = v.visit(n.left.plan)
 		n.right.plan = v.visit(n.right.plan)
 
 	case *interleavedJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "type", joinTypeStr(n.joinType))
-			v.observer.attr(name, "left table", formatTable(n.left.desc, n.left.index))
-		}
-		if v.observer.spans != nil {
-			v.observer.spans(name, "left spans", n.left.index, n.left.spans, n.left.hardLimit != 0)
-		}
-		if v.observer.expr != nil {
-			v.expr(name, "left filter", -1, n.leftFilter)
-		}
-		if v.observer.attr != nil {
-			v.observer.attr(name, "right table", formatTable(n.right.desc, n.right.index))
-		}
-		if v.observer.spans != nil {
-			v.observer.spans(name, "right spans", n.right.index, n.right.spans, n.right.hardLimit != 0)
-		}
-		if v.observer.expr != nil {
-			v.expr(name, "right filter", -1, n.rightFilter)
-			v.expr(name, "pred", -1, n.onCond)
-		}
-		if v.observer.attr != nil {
-			ancestor := "left"
-			lockingStrength, lockingWaitPolicy := n.left.lockingStrength, n.left.lockingWaitPolicy
-			if !n.leftIsAncestor {
-				ancestor = "right"
-				lockingStrength, lockingWaitPolicy = n.right.lockingStrength, n.right.lockingWaitPolicy
-			}
-			v.observer.attr(name, "ancestor", ancestor)
-
-			if lockingStrength != descpb.ScanLockingStrength_FOR_NONE {
-				v.observer.attr(name, "locking strength", lockingStrength.PrettyString())
-			}
-			if lockingWaitPolicy != descpb.ScanLockingWaitPolicy_BLOCK {
-				v.observer.attr(name, "locking wait policy", lockingWaitPolicy.PrettyString())
-			}
-		}
-
-		if v.observer.expr != nil {
-			v.expr(name, "pred", -1, n.onCond)
-		}
 
 	case *invertedFilterNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "inverted column", fmt.Sprintf("%d", n.invColumn+1))
-			v.observer.attr(name, "num spans", fmt.Sprintf("%d", len(n.expression.SpansToRead)))
-		}
 		n.input = v.visit(n.input)
 
 	case *invertedJoinNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", fmt.Sprintf("%s@%s", n.table.desc.Name, n.table.index.Name))
-			v.observer.attr(name, "type", joinTypeStr(n.joinType))
-		}
-		if v.observer.expr != nil {
-			v.expr(name, "", -1, n.invertedExpr)
-			if n.onExpr != nil && n.onExpr != tree.DBoolTrue {
-				v.expr(name, "onExpr", -1, n.onExpr)
-			}
-		}
-		if n.CanParallelize() {
-			v.observer.attr(name, "parallel", "")
-		}
 		n.input = v.visit(n.input)
 
 	case *limitNode:
-		if v.observer.expr != nil {
-			v.expr(name, "count", -1, n.countExpr)
-			v.expr(name, "offset", -1, n.offsetExpr)
-		}
 		n.plan = v.visit(n.plan)
 
 	case *max1RowNode:
 		n.plan = v.visit(n.plan)
 
 	case *distinctNode:
-		if v.observer.attr == nil {
-			n.plan = v.visit(n.plan)
-			break
-		}
-
-		if !n.distinctOnColIdxs.Empty() {
-			var buf bytes.Buffer
-			prefix := ""
-			columns := planColumns(n.plan)
-			n.distinctOnColIdxs.ForEach(func(col int) {
-				buf.WriteString(prefix)
-				buf.WriteString(columns[col].Name)
-				prefix = ", "
-			})
-			v.observer.attr(name, "distinct on", buf.String())
-			if n.nullsAreDistinct {
-				v.observer.attr(name, "nulls are distinct", "")
-			}
-			if n.errorOnDup != "" {
-				v.observer.attr(name, "error on duplicate", "")
-			}
-		}
-
-		if !n.columnsInOrder.Empty() {
-			var buf bytes.Buffer
-			prefix := ""
-			columns := planColumns(n.plan)
-			for i, ok := n.columnsInOrder.Next(0); ok; i, ok = n.columnsInOrder.Next(i + 1) {
-				buf.WriteString(prefix)
-				buf.WriteString(columns[i].Name)
-				prefix = ", "
-			}
-			v.observer.attr(name, "order key", buf.String())
-		}
-
 		n.plan = v.visit(n.plan)
 
 	case *sortNode:
-		if v.observer.attr != nil {
-			columns := planColumns(n.plan)
-			v.observer.attr(name, "order", n.ordering.String(columns))
-			if p := n.alreadyOrderedPrefix; p > 0 {
-				v.observer.attr(name, "already ordered", n.ordering[:p].String(columns))
-			}
-		}
 		n.plan = v.visit(n.plan)
 
 	case *groupNode:
-		if v.observer.attr != nil {
-			inputCols := planColumns(n.plan)
-			aggIdx := 0
-			for i, agg := range n.funcs {
-				if _, ok := n.aggIsGroupingColumn(i); ok {
-					// Don't show grouping columns as aggregations.
-					continue
-				}
-				var buf bytes.Buffer
-				if groupingCol, ok := n.aggIsGroupingColumn(i); ok {
-					buf.WriteString(inputCols[groupingCol].Name)
-				} else {
-					fmt.Fprintf(&buf, "%s(", agg.funcName)
-					if len(agg.argRenderIdxs) > 0 {
-						if agg.isDistinct() {
-							buf.WriteString("DISTINCT ")
-						}
-
-						for j, idx := range agg.argRenderIdxs {
-							if j != 0 {
-								buf.WriteString(", ")
-							}
-							buf.WriteString(inputCols[idx].Name)
-						}
-					}
-					buf.WriteByte(')')
-					if agg.filterRenderIdx != tree.NoColumnIdx {
-						fmt.Fprintf(&buf, " FILTER (WHERE %s)", inputCols[agg.filterRenderIdx].Name)
-					}
-				}
-				v.observer.attr(name, fmt.Sprintf("aggregate %d", aggIdx), buf.String())
-				aggIdx++
-			}
-			if len(n.groupCols) > 0 {
-				var cols []string
-				for _, c := range n.groupCols {
-					cols = append(cols, inputCols[c].Name)
-				}
-				v.observer.attr(name, "group by", strings.Join(cols, ", "))
-			}
-			if len(n.groupColOrdering) > 0 {
-				v.observer.attr(name, "ordered", n.groupColOrdering.String(inputCols))
-			}
-			if n.isScalar {
-				v.observer.attr(name, "scalar", "")
-			}
-		}
-
 		n.plan = v.visit(n.plan)
 
 	case *windowNode:
-		if v.observer.expr != nil {
-			for i, agg := range n.funcs {
-				v.metadataExpr(name, "window", i, agg.expr)
-			}
-			for i, rexpr := range n.windowRender {
-				v.metadataExpr(name, "render", i, rexpr)
-			}
-		}
 		n.plan = v.visit(n.plan)
 
 	case *unionNode:
@@ -548,122 +192,20 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 		n.rows = v.visit(n.rows)
 
 	case *insertNode, *insertFastPathNode:
-		var run *insertRun
-		if ins, ok := n.(*insertNode); ok {
-			run = &ins.run
-		} else {
-			run = &n.(*insertFastPathNode).run.insertRun
-		}
-
-		if v.observer.attr != nil {
-			var buf bytes.Buffer
-			buf.WriteString(run.ti.tableDesc().Name)
-			buf.WriteByte('(')
-			for i := range run.insertCols {
-				if i > 0 {
-					buf.WriteString(", ")
-				}
-				buf.WriteString(run.insertCols[i].Name)
-			}
-			buf.WriteByte(')')
-			v.observer.attr(name, "into", buf.String())
-			v.observer.attr(name, "strategy", run.ti.desc())
-			if run.ti.autoCommit == autoCommitEnabled {
-				v.observer.attr(name, "auto commit", "")
-			}
-		}
-
 		if ins, ok := n.(*insertNode); ok {
 			ins.source = v.visit(ins.source)
-		} else {
-			ins := n.(*insertFastPathNode)
-			if v.observer.attr != nil {
-				for i := range ins.run.fkChecks {
-					c := &ins.run.fkChecks[i]
-					tabDesc := c.ReferencedTable.(*optTable).desc
-					idxDesc := c.ReferencedIndex.(*optIndex).desc
-					v.observer.attr(name, "FK check", fmt.Sprintf("%s@%s", tabDesc.Name, idxDesc.Name))
-				}
-			}
-			if len(ins.input) != 0 {
-				if v.observer.attr != nil {
-					v.observer.attr(name, "size", formatValuesSize(len(ins.input), len(ins.input[0])))
-				}
-
-				if v.observer.expr != nil {
-					v.metadataTuples(name, ins.input)
-				}
-			}
 		}
 
 	case *upsertNode:
-		if v.observer.attr != nil {
-			var buf bytes.Buffer
-			buf.WriteString(n.run.tw.tableDesc().Name)
-			buf.WriteByte('(')
-			for i := range n.run.insertCols {
-				if i > 0 {
-					buf.WriteString(", ")
-				}
-				buf.WriteString(n.run.insertCols[i].Name)
-			}
-			buf.WriteByte(')')
-			v.observer.attr(name, "into", buf.String())
-			v.observer.attr(name, "strategy", n.run.tw.desc())
-			if n.run.tw.autoCommit == autoCommitEnabled {
-				v.observer.attr(name, "auto commit", "")
-			}
-		}
-
 		n.source = v.visit(n.source)
 
 	case *updateNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "table", n.run.tu.tableDesc().Name)
-			if len(n.run.tu.ru.UpdateCols) > 0 {
-				var buf bytes.Buffer
-				for i := range n.run.tu.ru.UpdateCols {
-					if i > 0 {
-						buf.WriteString(", ")
-					}
-					buf.WriteString(n.run.tu.ru.UpdateCols[i].Name)
-				}
-				v.observer.attr(name, "set", buf.String())
-			}
-			v.observer.attr(name, "strategy", n.run.tu.desc())
-			if n.run.tu.autoCommit == autoCommitEnabled {
-				v.observer.attr(name, "auto commit", "")
-			}
-		}
-		if v.observer.expr != nil {
-			for i, cexpr := range n.run.computeExprs {
-				v.metadataExpr(name, "computed", i, cexpr)
-			}
-		}
-		// An updater has no sub-expressions, so nothing special to do here.
 		n.source = v.visit(n.source)
 
 	case *deleteNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "from", n.run.td.tableDesc().Name)
-			v.observer.attr(name, "strategy", n.run.td.desc())
-			if n.run.td.autoCommit == autoCommitEnabled {
-				v.observer.attr(name, "auto commit", "")
-			}
-		}
-		// A deleter has no sub-expressions, so nothing special to do here.
 		n.source = v.visit(n.source)
 
 	case *deleteRangeNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "from", n.desc.Name)
-			if n.autoCommitEnabled {
-				v.observer.attr(name, "auto commit", "")
-			}
-		}
-		if v.observer.spans != nil {
-			v.observer.spans(name, "spans", &n.desc.PrimaryIndex, n.spans, false /* hardLimitSet */)
-		}
 
 	case *serializeNode:
 		v.visitConcrete(n.source)
@@ -677,29 +219,10 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 		}
 
 	case *createViewNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "query", n.viewQuery)
-		}
-
 	case *setVarNode:
-		if v.observer.expr != nil {
-			for i, texpr := range n.typedValues {
-				v.metadataExpr(name, "value", i, texpr)
-			}
-		}
-
 	case *setClusterSettingNode:
-		if v.observer.expr != nil && n.value != nil {
-			v.metadataExpr(name, "value", -1, n.value)
-		}
 
 	case *delayedNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "source", n.name)
-			if n.indexConstraint != nil {
-				v.observer.attr(name, "constraint", n.indexConstraint.String())
-			}
-		}
 		if n.plan != nil {
 			n.plan = v.visit(n.plan)
 		}
@@ -728,29 +251,13 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 		n.source = v.visit(n.source)
 
 	case *spoolNode:
-		if n.hardLimit > 0 && v.observer.attr != nil {
-			v.observer.attr(name, "limit", fmt.Sprintf("%d", n.hardLimit))
-		}
 		n.source = v.visit(n.source)
 
 	case *saveTableNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "target", n.target.String())
-		}
 		n.source = v.visit(n.source)
 
 	case *showTraceReplicaNode:
 		n.plan = v.visit(n.plan)
-
-	case *explainPlanNode:
-		// We check whether planNode is nil because the plan might be
-		// represented physically. We don't yet have a walker over such
-		// representation, so we simply short-circuit.
-		// TODO(yuzefovich): implement that walker and use it here.
-		if n.plan.main.planNode == nil {
-			return
-		}
-		n.plan.main.planNode = v.visit(n.plan.main.planNode)
 
 	case *cancelQueriesNode:
 		n.rows = v.visit(n.rows)
@@ -765,16 +272,8 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 		n.rows = v.visit(n.rows)
 
 	case *setZoneConfigNode:
-		if v.observer.expr != nil {
-			v.metadataExpr(name, "yaml", -1, n.yamlConfig)
-		}
 
 	case *projectSetNode:
-		if v.observer.expr != nil {
-			for i, texpr := range n.exprs {
-				v.metadataExpr(name, "render", i, texpr)
-			}
-		}
 		n.source = v.visit(n.source)
 
 	case *rowSourceToPlanNode:
@@ -786,78 +285,16 @@ func (v *planVisitor) visitInternal(plan planNode, name string) {
 		n.plan = v.visit(n.plan)
 
 	case *scanBufferNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "label", n.label)
-		}
 
 	case *bufferNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "label", n.label)
-		}
 		n.plan = v.visit(n.plan)
 
 	case *recursiveCTENode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "label", n.label)
-		}
 		n.initial = v.visit(n.initial)
 
 	case *exportNode:
-		if v.observer.attr != nil {
-			v.observer.attr(name, "destination", n.destination)
-		}
 		n.source = v.visit(n.source)
 	}
-}
-
-// expr wraps observer.expr() and provides it with the current node's
-// name.
-func (v *planVisitor) expr(nodeName string, fieldName string, n int, expr tree.Expr) {
-	if v.err != nil {
-		return
-	}
-	v.observer.expr(observeAlways, nodeName, fieldName, n, expr)
-}
-
-// metadataExpr wraps observer.expr() and provides it with the current node's
-// name, with verbosity = metadata.
-func (v *planVisitor) metadataExpr(nodeName string, fieldName string, n int, expr tree.Expr) {
-	if v.err != nil {
-		return
-	}
-	v.observer.expr(observeMetadata, nodeName, fieldName, n, expr)
-}
-
-// metadataTuples calls metadataExpr for each expression in a matrix.
-func (v *planVisitor) metadataTuples(nodeName string, tuples [][]tree.TypedExpr) {
-	for i := range tuples {
-		for j, expr := range tuples[i] {
-			var fieldName string
-			if v.observer.attr != nil {
-				fieldName = fmt.Sprintf("row %d, expr", i)
-			}
-			v.metadataExpr(nodeName, fieldName, j, expr)
-		}
-	}
-}
-
-// formatTable returns a string of the form "<table_name>@<index_name>", or
-// "<table_name>@<index_name> (partial index)" if the index is partial.
-func formatTable(desc *sqlbase.ImmutableTableDescriptor, index *descpb.IndexDescriptor) string {
-	partial := ""
-	if index.IsPartial() {
-		partial = " (partial index)"
-	}
-	return fmt.Sprintf("%s@%s%s", desc.Name, index.Name, partial)
-}
-
-// formatValuesSize returns a string of the form "5 columns, 1 row".
-func formatValuesSize(numRows, numCols int) string {
-	return fmt.Sprintf(
-		"%d column%s, %d row%s",
-		numCols, util.Pluralize(int64(numCols)),
-		numRows, util.Pluralize(int64(numRows)),
-	)
 }
 
 // nodeName returns the name of the given planNode as string.  The
@@ -892,24 +329,6 @@ func nodeName(plan planNode) string {
 	}
 
 	return name
-}
-
-func joinTypeStr(t descpb.JoinType) string {
-	switch t {
-	case descpb.InnerJoin:
-		return "inner"
-	case descpb.LeftOuterJoin:
-		return "left outer"
-	case descpb.RightOuterJoin:
-		return "right outer"
-	case descpb.FullOuterJoin:
-		return "full outer"
-	case descpb.LeftSemiJoin:
-		return "semi"
-	case descpb.LeftAntiJoin:
-		return "anti"
-	}
-	panic(errors.AssertionFailedf("unknown join type %s", t))
 }
 
 // planNodeNames is the mapping from node type to strings.  The
@@ -957,7 +376,6 @@ var planNodeNames = map[reflect.Type]string{
 	reflect.TypeOf(&dropViewNode{}):                "drop view",
 	reflect.TypeOf(&errorIfRowsNode{}):             "error if rows",
 	reflect.TypeOf(&explainDistSQLNode{}):          "explain distsql",
-	reflect.TypeOf(&explainPlanNode{}):             "explain plan",
 	reflect.TypeOf(&explainNewPlanNode{}):          "explain plan",
 	reflect.TypeOf(&explainVecNode{}):              "explain vectorized",
 	reflect.TypeOf(&exportNode{}):                  "export",
